# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HDD Painel API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/gates/{gate_id}": {
      "get": {
        "tags": [
          "gates"
        ],
        "summary": "Get Gate",
        "operationId": "get_gate_api_gates__gate_id__get",
        "parameters": [
          {
            "name": "gate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Gate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gates/{gate_id}/approve": {
      "post": {
        "tags": [
          "gates"
        ],
        "summary": "Approve Gate",
        "operationId": "approve_gate_api_gates__gate_id__approve_post",
        "parameters": [
          {
            "name": "gate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Gate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateDecisionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gates/{gate_id}/reject": {
      "post": {
        "tags": [
          "gates"
        ],
        "summary": "Reject Gate",
        "operationId": "reject_gate_api_gates__gate_id__reject_post",
        "parameters": [
          {
            "name": "gate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Gate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateDecisionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "GateDecisionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "GateDecisionOut"
      },
      "GateOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "wave_id": {
            "type": "string",
            "title": "Wave Id"
          },
          "gate_type": {
            "type": "string",
            "title": "Gate Type"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "wave_id",
          "gate_type",
          "reason",
          "status"
        ],
        "title": "GateOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}